Init other games apps
