+++ backend/games/games/battles/models.py
@@ -1,0 +1,6 @@
+from django.db import models
+
+
+class BattleRequest(models.Model):
+    initiator_id = models.PositiveIntegerField()
+    battle_case_id = models.PositiveIntegerField()
