+++ backend/games/games/common/serializers.py
@@ -1,0 +1,10 @@
+from rest_framework import serializers
+
+
+class UserFundsStateSerializer(serializers.Serializer):
+    id = serializers.IntegerField(min_value=0)
+    advantage = serializers.FloatField()
+
+
+class SiteFundsSerializer(serializers.Serializer):
+    site_active_funds_per_hour = serializers.FloatField(min_value=0)
