+++ backend/games/games/mines/api/urls.py
@@ -1,0 +1,1 @@
+
