+++ backend/games/games/mines/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class MinesConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'mines'
