+++ backend/games/games/upgrade/serializers.py
@@ -1,13 +1,6 @@
 from rest_framework import serializers
 
-
-class UserFundsStateSerializer(serializers.Serializer):
-    id = serializers.IntegerField(min_value=0)
-    advantage = serializers.FloatField()
-
-
-class SiteFundsSerializer(serializers.Serializer):
-    site_active_funds_per_hour = serializers.FloatField(min_value=0)
+from common.serializers import UserFundsStateSerializer, SiteFundsSerializer
 
 
 class UpgradeRequestSerializer(serializers.Serializer):
