+++ backend/products/products/games/api/endpoints/base.py
@@ -1,4 +1,3 @@
-from django.conf import settings
 from rest_framework.generics import RetrieveAPIView, CreateAPIView
