+++ backend/products/products/inventory/views.py
@@ -1,3 +1,1 @@
-from django.shortcuts import render
-
 # Create your views here.
